Adicionado: Missão 06

+++ src/main/java/br/ifpe/web2/missoes/service/CargoService.java
@@ -14,33 +14,44 @@ public class CargoService {
 
 	@Autowired private CargoDAO cargos;
 	
-	public Cargo save(Cargo cargo) {
+	@Autowired private FuncionarioService funcService;
+	
+	public Cargo salvar(Cargo cargo) {
 		return this.cargos.save(cargo);
 	}
-	public Optional<Cargo> findById(Integer id) {
+	
+	public Optional<Cargo> obterPorId(Integer id) {
 		return this.cargos.findById(id);
 	}
-	public List<Cargo> findAll() {
-		return this.cargos.findAll();
-	}
-	public List<Cargo> findFirst10ByOrderByDescricaoAsc() {
-		return this.cargos.findFirst10ByOrderByDescricaoAsc();
-	}
-	public List<Cargo> findAllByDescricao(String descricao) {
-		return this.cargos.findAllByDescricao(descricao);
-	}
-	public Optional<Cargo> findByDescricaoEquals(String descricao) {
+	public Optional<Cargo> obterPorDescricao(String descricao) {
 		return this.cargos.findByDescricaoEquals(descricao);
 	}
-	public void deleteById(Integer id) {
-		this.cargos.deleteById(id);
+	
+	public List<Cargo> listarTodos() {
+		return this.cargos.findAll();
 	}
-	public List<Cargo> findAllByAtivo(boolean ativo) {
-		return this.cargos.findAllByAtivo(ativo);
+	public List<Cargo> listarTodosAtivos() {
+		return this.cargos.findAllByAtivo(true);
+	}
+	public List<Cargo> listarPrimeiros10OrdenadosPorDescricao() {
+		return this.cargos.findFirst10ByOrderByDescricaoAsc();
+	}
+	public List<Cargo> listarTodosPorDescricao(String descricao) {
+		return this.cargos.findAllByDescricao(descricao);
 	}
 	
 	public boolean existe(Cargo cargo) {
 		return this.cargos.existsByDescricao(cargo.getDescricao());
+	}
+	
+	public void deletarPorId(Integer id) throws Exception {
+		Cargo cargo = this.cargos.findById(id).orElseThrow(() -> new Exception("Cargo inexistente."));
+		if (this.funcService.existeComCargo(cargo)) {
+			throw new Exception("Não é possível excluir o(s) cargo(s) de funcionário"
+					+ " selecionado(s) devido a vínculos com outras informações.");
+		}
+		
+		this.cargos.deleteById(id);
 	}
 	
 }
